Add risk_on and last_update state to config file.
The rules processor now uses the risk_on state from the config
file instead of the broker position of a ticker since it might
have multiple alias rules for the same ticker.

Also, the last_update field is used to prevent the risk_on state from
oscillating if the rebalancer needs to re-run due to any issues.